requirements.txt
boto3
pydo
python-dotenv